Implemented push and pull features

+++ scritps/pull.py
@@ -1,18 +1,24 @@
 from push import *
 from actions import *
-    if len(solidityCommits) == len(localCommits):
-        print("Your code is Already Ip To Date")
-    elif len(solidityCommits) < len(localCommits):
-         print("You need to push")
-    elif len(solidityCommits) > len(localCommits):
-         counter = len(localCommits)-1
-         for i in range(counter,len(solidityCommits)):
-             pullCommit(i,solidityCommits[i].authorName,solidityCommits[i].commitHash,
-             solidityCommits[i].date,solidityCommits[i].message,solidityCommits[i].addedFiles,solidityCommits[i].addLines,solidityCommits[i].removeFiles,solidityCommits[i].RemoveLines)
-           
- 
- 
- 
+
+# if len(solidityCommits) == len(Commits):
+#     print("Your code is Already Ip To Date")
+# elif len(solidityCommits) < len(Commits):
+#     print("You need to push")
+# elif len(solidityCommits) > len(Commits):
+#      counter = len(Commits)-1
+#      for i in range(counter,len(solidityCommits)):
+#          pullCommit(i,solidityCommits[i].authorName,solidityCommits[i].commitHash,
+#              solidityCommits[i].date,solidityCommits[i].message,solidityCommits[i].addedFiles,solidityCommits[i].addLines,solidityCommits[i].removeFiles,solidityCommits[i].RemoveLines)
+
+# info = ""
+# with open('info.json') as f:
+#     info = json.load(f)
+#     print(info)
+
+
+tempCommit = {list(commitsJson.keys())[-1] : commitsJson[list(commitsJson.keys())[-1]]}
+commitObj = tempCommit.get(list(tempCommit.keys())[0])
  
    
 def pullCommit(commitIndex,authorAddress,authorName,commitHash,date,message,addeFiles,addLines,removeFiles,RemoveLines):
@@ -28,19 +34,45 @@ def pullCommit(commitIndex,authorAddress,authorName,commitHash,date,message,adde
 
 def commitsAction(addedFiles,addedLines,removedFiles,removedLines):
     #create Files
-    for i in range(0,len(addedFiles)):
-        createFile(addedFiles[i],"")
-   
+    if(addedFiles):
+        for i in range(0,len(addedFiles) - 1):
+            createFile(addedFiles[i],"")
+
     #deleteFiles
-    for i in range(0,len(removedFiles))
-        deleteFile(removedFiles[i])     
-   
-   #add Lines
-   for i in range(0,len(addedLines)):
-       add = addedLines[i].split("@@@")
-       addOrdeleteLines(add[0],add[1],add[2])
-   
-   #remove Lines
-   for i in range(0,len(removedLines)):
-       remove = addedLines[i].split("@@@")
-       addOrdeleteLines(add[0],add[1],"")      
+    if(removedFiles):
+        for i in range(0,len(removedFiles) - 1):
+            deleteFile(removedFiles[i])     
+
+    if(removedLines):
+        for i in range(0,len(removedLines)):
+            tmpDict = removedLines[i]
+            location = tmpDict.get("File Location")
+            if(len(tmpDict.keys()) > 1):
+                for j in range(1, len(tmpDict.keys())):
+                    tmpDictLine = tmpDict.get(list(tmpDict.keys())[1])
+                    if(tmpDictLine != None):
+                        number = int(tmpDictLine.get("number"))
+                        content = tmpDictLine.get("content")
+                        deleteLines(location, number, content) 
+
+    #add Lines
+    if(addedLines):
+        for i in range(0,len(addedLines)):
+            tmpDict = addedLines[i]
+            location = tmpDict.get("File Location")
+            if(len(tmpDict.keys()) > 1):
+                for j in range(1, len(tmpDict.keys())):
+                    tmpDictLine = tmpDict.get(list(tmpDict.keys())[j])
+                    if(tmpDictLine != None):
+                        number = int(tmpDictLine.get("number"))
+                        content = tmpDictLine.get("content")
+                        addLines(location, number, content)  
+
+change = commitObj.get("change")
+commitsAction(change.get("files added"), change.get("Added Lines"), change.get("files deleted") , change.get("Removed Lines"))
+
+
+
+
+
+
